// apps/odonto-front/src/generated/ts/ClinicUserDto.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type ClinicUserDto = {
    /**
     * @type number
    */
    id: number;
    /**
     * @type string
    */
    name: string;
    /**
     * @type string
    */
    email: string;
    /**
     * @type string
    */
    role: string;
    /**
     * @type boolean
    */
    isActive: boolean;
    /**
     * @type string
    */
    avatarUrl: string | null;
};